refactor: 도메인이 dto 를 의존하지 않도록 변경 및 Piece 구하는 역할 State 로 이전
